movility_ai/requirements.txt
# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Google Agent Development Kit (ADK)
google-genai>=0.8.0

# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# Visualization and charting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
numpy>=1.24.0
pandas>=2.0.0
